optimise convert_rgb_to_yuyv in similar manner to alpha_blend
